# integrate isDeflate and isGzip