--- a/lib/utils.js
+++ b/lib/utils.js
@@ -6,17 +6,10 @@ function isDeflate (buffer) {
     typeof buffer === 'object' &&
     buffer !== null &&
     buffer.length > 1 &&
-    // CMF
-    (
-      buffer[0] === 0x78 ||
-      buffer[0] === 0x68 ||
-      buffer[0] === 0x58 ||
-      buffer[0] === 0x48 ||
-      buffer[0] === 0x38 ||
-      buffer[0] === 0x28 ||
-      buffer[0] === 0x18 ||
-      buffer[0] === 0x08
-    ) &&
+    // CM = 8 denotes the "deflate" compression method
+    (buffer[0] & 0x0f) === 0x08 &&
+    // CINFO Values of above 7 are not allowed by RFC 1950
+    (buffer[0] >> 0x04) < 0x08 &&
     // The FCHECK value must be such that CMF and FLG, when viewed as
     // a 16-bit unsigned integer stored in MSB order (CMF*256 + FLG),
     // is a multiple of 31.