--- a/lib/utils.js
+++ b/lib/utils.js
@@ -9,7 +9,7 @@ function isDeflate (buffer) {
     // CM = 8 denotes the "deflate" compression method
     (buffer[0] & 0x0f) === 0x08 &&
     // CINFO Values of above 7 are not allowed by RFC 1950
-    (buffer[0] >> 0x04) < 0x08 &&
+    (buffer[0] & 0x80) === 0 &&
     // The FCHECK value must be such that CMF and FLG, when viewed as
     // a 16-bit unsigned integer stored in MSB order (CMF*256 + FLG),
     // is a multiple of 31.