--- a/package.json
+++ b/package.json
@@ -8,8 +8,6 @@
     "encoding-negotiator": "^2.0.1",
     "fastify-plugin": "^3.0.0",
     "into-stream": "^6.0.0",
-    "is-deflate": "^1.0.0",
-    "is-gzip": "^2.0.0",
     "is-zip": "^1.0.0",
     "mime-db": "^1.51.0",
     "minipass": "^3.1.6",