--- a/test/utils.test.js
+++ b/test/utils.test.js
@@ -4,7 +4,7 @@ const { createReadStream } = require('fs')
 const { Socket } = require('net')
 const { Duplex, PassThrough, Readable, Stream, Transform, Writable } = require('stream')
 const { test } = require('tap')
-const { isStream } = require('../lib/utils')
+const { isStream, isDeflate, isGzip } = require('../lib/utils')
 
 test('isStream() utility should be able to detect Streams', async (t) => {
   t.plan(12)
@@ -24,4 +24,37 @@ test('isStream() utility should be able to detect Streams', async (t) => {
   t.equal(isStream(null), false)
   t.equal(isStream(undefined), false)
   t.equal(isStream(''), false)
+})
+
+test('isDeflate() utility should be able to detect deflate compressed Buffer', async (t) => {
+  t.plan(11)
+
+  t.equal(isDeflate(Buffer.alloc(0)), false)
+  t.equal(isDeflate(Buffer.alloc(0)), false)
+  t.equal(isDeflate(Buffer.from([0x78])), false)
+  t.equal(isDeflate(Buffer.from([0x78, 0x00])), false)
+  t.equal(isDeflate(Buffer.from([0x78, 0x01])), true)
+  t.equal(isDeflate(Buffer.from([0x78, 0x9c])), true)
+  t.equal(isDeflate(Buffer.from([0x78, 0xda])), true)
+
+  t.equal(isDeflate({}), false)
+  t.equal(isDeflate(null), false)
+  t.equal(isDeflate(undefined), false)
+  t.equal(isDeflate(''), false)
+})
+
+test('isGzip() utility should be able to detect gzip compressed Buffer', async (t) => {
+  t.plan(10)
+
+  t.equal(isGzip(Buffer.alloc(0)), false)
+  t.equal(isGzip(Buffer.alloc(1)), false)
+  t.equal(isGzip(Buffer.alloc(2)), false)
+  t.equal(isGzip(Buffer.from([0x1f, 0x8b])), false)
+  t.equal(isGzip(Buffer.from([0x1f, 0x8b, 0x00])), false)
+  t.equal(isGzip(Buffer.from([0x1f, 0x8b, 0x08])), true)
+
+  t.equal(isGzip({}), false)
+  t.equal(isGzip(null), false)
+  t.equal(isGzip(undefined), false)
+  t.equal(isGzip(''), false)
 })