@Uzlopak 2022-06-13
@kibertoad 
@Fdawgs 

I have to ask you to again review it. On a smoking break I realized that i can do the CMF header check in isDeflate with bit operations. 

before:
```
valid x 223,947,412 ops/sec ±0.70% (90 runs sampled)
invalid x 831,593,213 ops/sec ±0.88% (93 runs sampled)
```

after:
```
valid x 1,027,199,609 ops/sec ±0.74% (89 runs sampled)
invalid x 1,038,269,211 ops/sec ±0.94% (91 runs sampled)
```

```
'use strict'

const { Suite } = require('benchmark')
const isDeflate = require('./lib/utils').isDeflate

const suite = new Suite()

const fakeDeflate = Buffer.from([0x78, 0x9d])
const fakeGzip = Buffer.from(Buffer.from([0x1f, 0x8b, 0x08]))

suite.add('valid', function () {
  isDeflate(fakeDeflate)
}).add('invalid', function () {
  isDeflate(fakeGzip)
})
suite
  .on('cycle', function (event) {
    console.log(String(event.target))
  })
  .on('complete', function () {
    console.log('Fastest is ' + this.filter('fastest').map('name'))
    console.log('Slowest is ' + this.filter('slowest').map('name'))
  })
  .run({ async: true })

```